fix(cli): address diagnostic logging review findings
Route the remaining PTY diagnostic through -E, keep remote stderr tests private, and set evidence-backed OpenSSH compatibility floors from complete Linux and macOS runs.
