+++ tests/log_file_test.rs
@@ -16,7 +16,6 @@ use std::path::{Path, PathBuf};
 use std::process::{Command, Output};
 
 use bssh::cli::Cli;
-use bssh::executor::NodeOutputWriter;
 use clap::Parser;
 use tempfile::tempdir;
 
@@ -84,46 +83,6 @@ fn human_diagnostics_are_routed_and_repeated_runs_append() {
             .count(),
         2
     );
-}
-
-#[test]
-fn remote_stderr_helper() {
-    if std::env::var_os("BSSH_LOG_FILE_TEST_HELPER").is_none() {
-        return;
-    }
-
-    let log = PathBuf::from(
-        std::env::var_os("BSSH_LOG_FILE_TEST_PATH").expect("helper log path should be set"),
-    );
-    bssh::utils::diagnostics::set_log_file(&log).expect("helper should open diagnostic log");
-    bssh::diagnosticln!("bssh-owned diagnostic");
-    NodeOutputWriter::new_with_no_prefix("example", true)
-        .write_stderr_lines("remote command stderr")
-        .expect("helper should write remote stderr");
-}
-
-#[test]
-fn remote_command_stderr_stays_on_fd_2() {
-    let directory = tempdir().expect("failed to create temporary directory");
-    let log = directory.path().join("bssh.log");
-    let output = Command::new(std::env::current_exe().expect("test executable should exist"))
-        .args(["--exact", "remote_stderr_helper", "--nocapture"])
-        .env("BSSH_LOG_FILE_TEST_HELPER", "1")
-        .env("BSSH_LOG_FILE_TEST_PATH", &log)
-        .output()
-        .expect("failed to run remote stderr helper");
-
-    assert!(output.status.success(), "remote stderr helper should pass");
-    let stderr = String::from_utf8_lossy(&output.stderr);
-    assert!(
-        stderr.contains("remote command stderr"),
-        "stderr was: {stderr}"
-    );
-    assert!(!stderr.contains("bssh-owned diagnostic"));
-
-    let contents = std::fs::read_to_string(&log).expect("failed to read diagnostic log");
-    assert!(contents.contains("bssh-owned diagnostic"));
-    assert!(!contents.contains("remote command stderr"));
 }
 
 #[test]
